requirements.txt
Django==3.2.12
mysqlclient
django-simpleui
django-import-export